pkp/pkp-lib#4787 moved migrations for 3.5.0
